Checkpoint: updating index and cache infrastructure

+++ 201102_parser_updates1/plugins/net.sf.sveditor.core/src/net/sf/sveditor/core/db/index/cache/SVDBDirFS.java
@@ -6,6 +6,8 @@ import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
+import java.io.RandomAccessFile;
+import java.util.List;
 
 public class SVDBDirFS implements ISVDBFS {
 	private File				fDBDir;
@@ -13,14 +15,42 @@ public class SVDBDirFS implements ISVDBFS {
 	public SVDBDirFS(File root) {
 		fDBDir = root;
 	}
+	
+	public String getRoot() {
+		return fDBDir.getAbsolutePath();
+	}
+	
+	public void removeStoragePath(List<File> db_file_list) {
+		db_file_list.remove(fDBDir);
+	}
 
 	public InputStream openFileRead(String path) {
 		InputStream ret = null;
 		try {
 			ret = new FileInputStream(new File(fDBDir, path));
+			// ret = new MappedByteBufferInputStream(new File(fDBDir, path));
 		} catch (IOException e) {}
 		
 		return ret;
+	}
+	
+	public RandomAccessFile openChannelRead(String path) {
+		RandomAccessFile ret = null;
+		File target = new File(fDBDir, path);
+		
+		try {
+			ret = new RandomAccessFile(target, "r");
+		} catch (IOException e) {}
+		
+		return ret;
+	}
+	
+	public void closeChannel(RandomAccessFile ch) {
+		try {
+			ch.close();
+		} catch (IOException e) {
+			e.printStackTrace();
+		}
 	}
 
 	public OutputStream openFileWrite(String path) {
@@ -37,6 +67,23 @@ public class SVDBDirFS implements ISVDBFS {
 		return ret;
 	}
 	
+	public RandomAccessFile openChannelWrite(String path) {
+		RandomAccessFile ret = null;
+		File target = new File(fDBDir, path);
+		File target_p = target.getParentFile();
+		
+		if (!target_p.exists()) {
+			target_p.mkdirs();
+		}
+		
+		try {
+			ret = new RandomAccessFile(new File(fDBDir, path), "rw");
+			ret.setLength(0);
+		} catch (IOException e) {}
+		
+		return ret;
+	}
+	
 	public void close(InputStream in) {
 		try {
 			in.close();
@@ -45,7 +92,11 @@ public class SVDBDirFS implements ISVDBFS {
 
 	public boolean fileExists(String path) {
 		File file = new File(fDBDir, path);
-		return file.exists();
+		if (file.exists()) {
+			return true;
+		} else {
+			return false;
+		}
 	}
 	
 	public long lastModified(String path) {
@@ -57,10 +108,18 @@ public class SVDBDirFS implements ISVDBFS {
 	public void delete(String path) {
 		if (path.equals("")) {
 			if (fDBDir.exists()) {
+				System.out.println("Delete root dir " + fDBDir.getAbsolutePath());
+				try {
+					throw new Exception();
+				} catch (Exception e) {
+					e.printStackTrace();
+				}
 				delete_tree(fDBDir);
 			}
 		} else {
 			File file = new File(fDBDir, path);
+
+			System.out.println("Delete \"" + file.getAbsolutePath() + "\"");
 
 			if (file.isDirectory()) {
 				delete_tree(file);
@@ -79,22 +138,30 @@ public class SVDBDirFS implements ISVDBFS {
 	}
 	
 	private void delete_tree(File p) {
-		for (File f : p.listFiles()) {
-			if (f.getName().equals("..") || f.getName().equals(".")) {
-				System.out.println("[ERROR] " + f.getName());
-				continue;
-			}
-			if (f.isDirectory()) {
-				delete_tree(f);
-			}
-		}
-		for (File f : p.listFiles()) {
-			if (f.getName().equals("..") || f.getName().equals(".")) {
-				System.out.println("[ERROR] " + f.getName());
-				continue;
-			}
-			if (f.isFile()) {
-				f.delete();
+		if (p.isFile()) {
+			p.delete();
+		} else {
+			if (p.exists()) {
+				for (File f : p.listFiles()) {
+					if (f.getName().equals("..") || f.getName().equals(".")) {
+						System.out.println("[ERROR] " + f.getName());
+						continue;
+					}
+					if (f.isDirectory()) {
+						delete_tree(f);
+					}
+				}
+				for (File f : p.listFiles()) {
+					if (f.getName().equals("..") || f.getName().equals(".")) {
+						System.out.println("[ERROR] " + f.getName());
+						continue;
+					}
+					if (f.isFile()) {
+						f.delete();
+					}
+				}
+
+				p.delete();
 			}
 		}
 	}
